Moving FormInput, FormSelect, and PageLoading into src/payment module.
… in appropriate places to their consumers.  The form input and select are in payment-form because that’s where they’re used.  PageLoading is up top.
