+++ src/payment/checkout/payment-form/CardHolderInformation.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux';
 import { clearFields, Field } from 'redux-form';
 import { injectIntl, intlShape, FormattedMessage } from '@edx/frontend-i18n';
 
-import FormInput from '../../../common/components/FormInput';
-import FormSelect from '../../../common/components/FormSelect';
+import FormInput from './FormInput';
+import FormSelect from './FormSelect';
 
 import { countryOptionsSelector } from '../../data/selectors';
 import messages from './CardHolderInformation.messages';
