+++ src/payment/checkout/payment-form/StateProvinceFormInput.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux';
 import { Field } from 'redux-form';
 import { injectIntl, intlShape, FormattedMessage } from '@edx/frontend-i18n';
 
-import FormInput from '../../../common/components/FormInput';
-import FormSelect from '../../../common/components/FormSelect';
+import FormInput from './FormInput';
+import FormSelect from './FormSelect';
 import getStates from './utils/countryStatesMap';
 import messages from './StateProvinceFormInput.messages';
